Amounts in simple float format

+++ accounting/line_item.go
@@ -12,7 +12,7 @@ type LineItem struct {
 	Quantity float64 `json:"Quantity,omitempty" xml:"Quantity,omitempty"`
 
 	// LineItem Unit Amount
-	UnitAmount float64 `json:"UnitAmount" xml:"UnitAmount"`
+	UnitAmount Float `json:"UnitAmount" xml:"UnitAmount"`
 
 	// See Items
 	ItemCode string `json:"ItemCode,omitempty" xml:"ItemCode,omitempty"`
@@ -24,10 +24,10 @@ type LineItem struct {
 	TaxType string `json:"TaxType,omitempty" xml:"TaxType,omitempty"`
 
 	// The tax amount is auto calculated as a percentage of the line amount (see below) based on the tax rate. This value can be overriden if the calculated <TaxAmount> is not correct.
-	TaxAmount float64 `json:"TaxAmount,omitempty" xml:"TaxAmount,omitempty"`
+	TaxAmount Float `json:"TaxAmount,omitempty" xml:"TaxAmount,omitempty"`
 
 	// If you wish to omit either of the <Quantity> or <UnitAmount> you can provide a LineAmount and Xero will calculate the missing amount for you. The line amount reflects the discounted price if a DiscountRate has been used . i.e LineAmount = Quantity * Unit Amount * ((100 – DiscountRate)/100)
-	LineAmount float64 `json:"LineAmount,omitempty" xml:"LineAmount,omitempty"`
+	LineAmount Float `json:"LineAmount,omitempty" xml:"LineAmount,omitempty"`
 
 	// Optional Tracking Category – see Tracking.  Any LineItem can have a maximum of 2 <TrackingCategory> elements.
 	Tracking []TrackingCategory `json:"Tracking,omitempty" xml:"Tracking>TrackingCategory,omitempty"`
